build basic JIRA container
We start by setting up our own JIRA useing a PostgresSQL database. I run
the setup wizzard and anything looks like it is working. We should
ensure our volumes are persistent so we do not have to rerun the setup
anytime we start up.

I use networking to connect the container in place of links. This makes
it possible to reload container on update besides some other neat stuff.
As this feature is still considered experimentel I need to execute
docker-compose with a param. To make this easier I created a startup
script.